Rename package to windowstate
