+++ DualView/src/main/java/com/microsoft/device/display/samples/dualview/MainActivity.kt
@@ -12,8 +12,8 @@ import androidx.lifecycle.ViewModelProvider
 import com.microsoft.device.display.samples.dualview.models.AppStateViewModel
 import com.microsoft.device.display.samples.dualview.ui.home.DualViewApp
 import com.microsoft.device.display.samples.dualview.ui.theme.DualViewComposeSampleTheme
-import com.microsoft.device.dualscreen.window.WindowState
-import com.microsoft.device.dualscreen.window.rememberWindowState
+import com.microsoft.device.dualscreen.windowstate.WindowState
+import com.microsoft.device.dualscreen.windowstate.rememberWindowState
 
 class MainActivity : AppCompatActivity() {
     private lateinit var windowState: WindowState
