+++ WindowState/src/main/java/com/microsoft/device/dualscreen/windowstate/WindowMode.kt
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-package com.microsoft.device.dualscreen.window
+package com.microsoft.device.dualscreen.windowstate
 
 enum class WindowMode {
     SINGLE_PORTRAIT,
