+++ WindowState/src/main/java/com/microsoft/device/dualscreen/windowstate/WindowSizeClass.kt
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-package com.microsoft.device.dualscreen.window
+package com.microsoft.device.dualscreen.windowstate
 
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
